Theoretical Assignment
Description of the function f as a truth table and as a Boolean formula.


We will use the following secure two-party computation scenario as a running example for all the assignments: 

Alice has blood type x, Bob has blood type y. We want Alice to learn z=f(x,y), where f is the function that outputs 1 if a person with blood type x can receive a blood transfusion from a person with blood type y (and 0 otherwise). Bob should learn nothing.